requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
insightface>=0.7.3
onnxruntime>=1.15.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Scientific Computing
scipy>=1.11.0

# Metadata Handling
iptcinfo3>=2.1.0
